Using recursion to traverse through trees
Problem Description: https://leetcode.com/problems/path-sum/description/
Using recursion to traverse through a tree and having an extra variable sum that we subtract from to keep track of current count. If we arrive at a leaf then we check if the current value at leaf node is equal to the leftover sum.